Feature/refactor server router (#290)
* refactoring server

* more refactoring

* fix lint

* moved router into its own folder. moved helpers into router folder. split out more helpers in router into helpers.

* fix conflicts

* moving code into separate files for route handlers

* fix lint
