+++ server/router/routes/tasklist-handler.js
@@ -19,8 +19,23 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-const momentToLong = require('./moment-to-long');
+const tasklistHandler = async ctx => {
+  const { domain, taskListName } = ctx.params;
+  const descTaskList = async taskListType =>
+    await ctx.cadence.describeTaskList({
+      domain,
+      taskList: { name: taskListName },
+      taskListType,
+    });
 
-module.exports = {
-  momentToLong,
+  const activityList = await descTaskList('Activity');
+  const decisionList = await descTaskList('Decision');
+  const activityPollerList = activityList.pollers || [];
+  const decisionPollerList = decisionList.pollers || [];
+
+  const taskList = { pollers: [...activityPollerList, ...decisionPollerList] };
+
+  ctx.body = taskList;
 };
+
+module.exports = tasklistHandler;
