+++ server/router/routes/workflow-archived-list-handler.js
@@ -1,0 +1,47 @@
+// Copyright (c) 2021 Uber Technologies Inc.
+//
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+const moment = require('moment');
+const { buildQueryString } = require('../helpers');
+
+const workflowArchivedListHandler = async ctx => {
+  const { nextPageToken, ...query } = ctx.query || {};
+  let queryString;
+
+  if (query.queryString) {
+    queryString = query.queryString;
+  } else {
+    const startTime = moment(query.startTime || NaN);
+    const endTime = moment(query.endTime || NaN);
+
+    ctx.assert(startTime.isValid() && endTime.isValid(), 400);
+    queryString = buildQueryString(startTime, endTime, query);
+  }
+
+  ctx.body = await ctx.cadence.archivedWorkflows({
+    query: queryString,
+    nextPageToken: nextPageToken
+      ? Buffer.from(nextPageToken, 'base64')
+      : undefined,
+  });
+};
+
+module.exports = workflowArchivedListHandler;
