+++ server/router/routes/workflow-export-handler.js
@@ -1,0 +1,46 @@
+// Copyright (c) 2021 Uber Technologies Inc.
+//
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+const losslessJSON = require('lossless-json');
+
+const workflowExportHandler = async ctx => {
+  let nextPageToken;
+
+  do {
+    const page = await ctx.cadence.exportHistory({ nextPageToken });
+
+    if (!nextPageToken) {
+      ctx.status = 200;
+    }
+
+    ctx.res.write(
+      (nextPageToken ? ',' : '[') +
+        page.history.events.map(losslessJSON.stringify).join(',')
+    );
+    nextPageToken =
+      page.nextPageToken && Buffer.from(page.nextPageToken, 'base64');
+  } while (nextPageToken);
+
+  ctx.res.write(']');
+  ctx.body = '';
+};
+
+module.exports = workflowExportHandler;
